ブロックの当たり判定
当たり判定をより正しい挙動に修正した
プレイヤーがブロックの角から降りるとｶｸｯと変な挙動をとるためそこの修正をする必要がある
ジャンプを実装した

+++ ActionGame/ActionGame/BaseObject.cpp
@@ -2,7 +2,7 @@
 
 
 BaseObject::BaseObject(BaseScene* _baseScene, F_Vec2 _pos, F_Vec2 _vel,
-	int _top, int _under, int _left, int _right,
+	float _top, float _under, float _left, float _right,
 	int _tag)
 	:position{ _pos }
 	, baseScene{ _baseScene }
