+++ ActionGame/ActionGame/BaseObject.h
@@ -11,7 +11,7 @@ class BaseObject
 {
 public:
 	BaseObject(BaseScene* _baseScene,F_Vec2 _pos, F_Vec2 _vel,
-		int _top,int _under,int _left,int _right,
+		float _top,float _under,float _left,float _right,
 		int _tag);
 	~BaseObject();
 	virtual void Start() {}
